Add config init flow

+++ jmore-blog/src/main/java/com/whatakitty/jmore/blog/infrastructure/repository/mybatis/user/UserDO.java
@@ -12,7 +12,7 @@ public class UserDO {
 
     private Date gmtModified;
 
-    private Byte userType;
+    private Integer userType;
 
     private String username;
